ezyago-main/src/backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Settings:
    # --- Temel Ayarlar ---
    API_KEY: str = os.getenv("BINANCE_API_KEY")
    API_SECRET: str = os.getenv("BINANCE_API_SECRET")
    ENVIRONMENT: str = os.getenv("ENVIRONMENT", "LIVE")
    BOT_USERNAME: str = os.getenv("BOT_USERNAME", "admin")
    BOT_PASSWORD: str = os.getenv("BOT_PASSWORD", "changeme123")
    BASE_URL = "https://fapi.binance.com" if os.getenv("ENVIRONMENT", "TEST") == "LIVE" else "https://testnet.binancefuture.com"
    WEBSOCKET_URL = "wss://fstream.binance.com" if os.getenv("ENVIRONMENT", "TEST") == "LIVE" else "wss://stream.binancefuture.com"

    # --- İşlem Parametreleri ---
    LEVERAGE: int = 10
    ORDER_SIZE_USDT: float = 25.0
    TIMEFRAME: str = "15m"
    
    # --- Kâr/Zarar Ayarları (Sadece Stop Loss) ---
    # Bu stratejide kâr, bir sonraki ters sinyalde alınır.
    # Bu SL, sadece ani ve büyük ters hareketlere karşı bir sigortadır.
    STOP_LOSS_PERCENT: float = 0.04  # %1 Zarar Durdur

settings = Settings()
