ezyago-main/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-binance==1.0.19
firebase-admin==6.2.0
cryptography>=41.0.0,<46.0.0
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
pydantic==2.5.0
email-validator==2.1.0
httpx==0.25.2
websockets==12.0
